@@ -7,7 +7,6 @@ namespace Laminas\HttpHandlerRunner\Exception;
 use InvalidArgumentException;
 use Laminas\HttpHandlerRunner\Emitter;
 
-use function get_class;
 use function gettype;
 use function is_object;
 use function sprintf;
@@ -23,7 +22,7 @@ class InvalidEmitterException extends InvalidArgumentException implements Except
             '%s can only compose %s implementations; received %s',
             Emitter\EmitterStack::class,
             Emitter\EmitterInterface::class,
-            is_object($emitter) ? get_class($emitter) : gettype($emitter)
+            is_object($emitter) ? $emitter::class : gettype($emitter)
         ));
     }
 }
